requirements.txt
pyspark==3.2.0
kafka-python==2.0.2
py4j==0.10.9.3